+++ test/libsolidity/syntaxTests/structs/recursion/recursive_struct_as_contract_function_parameter.sol
@@ -1,0 +1,10 @@
+contract Test {
+    struct MyStructName {
+        address addr;
+        MyStructName[] x;
+    }
+
+    function f(MyStructName memory s) public {}
+}
+// ----
+// TypeError: (112-133): Internal or recursive type is not allowed for public or external functions.